[VirtualCluster] vcluster cli command support minimal install
